+++ c_jmpl/include/table.h
@@ -22,7 +22,8 @@ bool tableSet(GC* gc, Table* table, ObjString* key, Value value);
 bool tableDelete(Table* table, ObjString* key);
 void tableAddAll(GC* gc, Table* from, Table* to);
 
-ObjString* tableFindString(Table* table, const unsigned char* chars, int length, uint32_t hash);
+ObjString* tableFindString(Table* table, const unsigned char* chars, int length, uint64_t hash);
+Entry* tableFindJoinedStrings(GC* gc, Table* table, const unsigned char* a, int aLen, const unsigned char* b, int bLen, uint64_t hash);
 void tableRemoveWhite(Table* table);
 void markTable(GC* gc, Table* table);
 void tableDebugStats(Table* table);